requirements.txt
Django==3.1.3
environs==9.3.0
python-dotenv==0.10.3
django-admin-sortable2==0.7.7
django-tinymce==3.2.0
